Backend/requirements.txt
# ─────────────────────────────────────────────────────────────
# Analysis_CV_V2 — Python Dependencies
# ─────────────────────────────────────────────────────────────
# Install:  pip install -r requirements.txt

# --- Web Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12       # Required for FastAPI file uploads

# --- LangChain Core ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# --- LangChain Model Providers ---
langchain-anthropic>=0.3.0     # ChatAnthropic (Claude Sonnet)
langchain-openai>=0.3.0        # OpenAIEmbeddings

# --- LangGraph (Agent framework) ---
langgraph>=0.2.0

# --- Vector Store ---
faiss-cpu>=1.9.0               # Facebook AI Similarity Search

# --- Document Parsing ---
pdfplumber>=0.11.0             # PDF text extraction
python-docx>=1.1.0             # DOCX text extraction

# --- Configuration & Security ---
python-dotenv>=1.0.0           # Load .env files (never hardcode secrets!)
pydantic>=2.9.0                # Data validation & schemas

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0                  # Async HTTP client for FastAPI testing
